Impulse-Nutrition/shopify_mcp/pyproject.toml
[project]
name = "shopify-mcp"
version = "0.1.0"
description = "MCP server for Shopify store management (draft orders, etc.)"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.9.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
